app/requirements.txt
amqp==2.4.1
billiard==3.5.0.5
celery==4.2.1
Django==2.1.7
djangorestframework==3.9.1
coreapi==2.3.3
kombu==4.3.0
pytz==2018.9
vine==1.2.0
